refactor env c api
